Implements two new command 'sources' and 'kill-sources'. The first one reports the list of Information Sources (and its individual status) of a running server.  The 2nd one allows to kill remotely an Information Source.
It also re-implements the command line parsing mechanism. Since it
does not use getopt() any longer, it can use a much more flexible
parsing strategy that allows to place the main command and its
parameters at the end of the command line:

  cherokee-tweak [options] URL info
  cherokee-tweak [options] URL sources
  cherokee-tweak [options] URL kill-source ID
  cherokee-tweak [options] URL logrotate   PATH
  cherokee-tweak [options] URL trace       TRACE

The documentation does not reflect these changes yet.

git-svn-id: svn://cherokee-project.com/cherokee/trunk@5392 5dc97367-97f1-0310-9951-d761b3857238