Implements/Ports; "Feature Request: Server Tokens" issue 438.
I took the liberty to port the proposed patch and extend its functionality beyond what is published to the outside. And including what is published to foreign
scripts and proxy hosts. The point is obviously be secure by default, the point of hiding it: make it a little bit harder to guess what is actually 'the default'.

http://code.google.com/p/cherokee/issues/detail?id=438