handler_proxy: Don't set the 'got all' flag when POST data is received
That flag only use is to know if the connection can be re-used. Getting
all the POST data from the client has _nothing_ to do with that, we
must only set it when the response from the backend is complete.

Failure to do this results in the connection trying to be reused when
it's not ready ... In case of failure it's not closed and the next
request might try to use it whatever state its in.

Signed-off-by: Sylvain Munaut <s.munaut@whatever-company.com>