Fix #890
This patch addresses the problem described in bug #890.
Sadly the solution works around an architectural issue.
When parsing the headers, Cherokee checks if a handler
supports Content Range. Typically CGI doesn't but when
in the reply X-Sendfile is found the option is set,
based on the fact that handler_file does.

The latter is too late, because the header is already
parsed, since the parsing didn't set the range,
handler_file is unaware and doesn't set it. This first
commit is to show and solve the problem.