Fix #880
It is a good thing Cherokee allows to clone rules. Currently
cloning the default rule, creates a duplicate where the match
section cannot be changed. Obviously this is not expected.
This improvement fakes a default behavior by the request match
with regular expression /.*

Ideally we might need an any match, which helps with a global
configuration option implemented by a non-final, any match.
Think of the use case of protecting the entire virtual server
with a password.