ssl: Allow to specify a fixed length
By default openssl will only get 512 or 1024 (depending only of the
"export" setting). But 1024 is a bit short nowadays and 2048 is
recommended. It does break compatibility in very rare case, so it's
left by default to auto. (there is no negotiation, the client has to
support whatever the server sends)

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>