Implement a way to emit warnings as suggested by @Borkason in #993
A validation can validate, warn and give an error.
When we do not return any value the validation will pass.
Until we find an error, each warning will be added gracefully.
If we do in fact find one error, the for loop will break, and all
previously added warning and the last error is treated as error.

To show the "warnings" in the interface, we merge the dictionary
that we expect to return with our warning, this will override the
return value.

I consider this a hack, but for now it functions gracefully.