When X-Sendfile or X-Accel-Redirect is present we should not invoke the encoder.
When an X-Sendfile header is send, we should not try to encode
this file. In all likelihood the developer knows what happens,
and might even have added Content-Length to the connection.

The parsing of the X-Sendfile and X-Accel-Redirect is pulled before
the parsing of Content-Length because these headers can change the
result of cherokee_connection_should_include_length. This prevents
parsing the CGI header in a different sequence having different
results.

Fix #846