Prevent resolving to a unix socket from the proxy
Currently Issue #723 mentions that the proxy crashes when a
unix socket backend is selected. This is not the case anymore.
But I did observe chat because the length of host is null, the
resolver attempts (null):80. For now, bail out early, we can
later add the option to proxy over unix sockets.