Reverse proxy skips webdirectory in the request. Fix #1037
The use of a 'Directory' rule (or webdir) was is most cases
transparent. The backend was unaware that the request was
made via an extra path, and therefore it didn't matter if
you set up your favorite CMS under /cms or as the default
rule.

The proxy behaved differently. Every request was forwarded
as it came in. In the directory rule /proxy this would result
in: /proxy/request -> /proxy/request

Obviously this would not give a bad situation, if only you
could rewrite this "/proxy" from the original request.
Sadly you could only rewrite /request.

The proxy is now removes the webdir from the request.
If you would like to have the old behavior you can still
use a regular expression or a fullpath.