Don't set X-Forwarded-SSL: ON when headers indicate SSL was OFF before.
My guess is that this could be merged. The following behavior is changed:

When an incomming SSL connection has an X-Forwarded-SSL which is OFF
the X-Forwarded-SSL will never set to ON. The only way that X-Forwarded-SSL
can be ON:

 - there is a secure connection and no X-Forwarded-SSL header is present -> ON
 - there is a secure connection and X-Forwarded-SSL is ON -> ON

In all other cases X-Forwarded SSL will be set OFF:
 - there is no secure connection and no header X-Forward-SSL -> OFF
 - there is no secure connection and X-Forwarded-SSL is ON -> OFF

There is a potential optimisation to copy the existing off header.
For readibility I did not go in that direction.

Fix #1073