Change the cherokee_ ## klass ## _free to return void.
The idea of these modifications is - for now - not to be added into
master. I am interested if these changes can help me to find out
which function do need return validation and which don't.

While adding __attribute__((warn_unused_result)) to every
function out there is quite trivial, I am more interested
in where Cherokee is currently checking for problems,
where it should, and where it is plain useless to do so.

Some of the code changes in this commit did contain
return ret_error, but was never used, or caused problems
further down the road aka leaking memory.