When the server can't change the interpreter the specified user or group, the fork will be terminated.
Warning; This commit changes a fundamental function of Cherokee.

When you have specified a user/group pair, and the server cannot fork
to this specified user and/or group, the process wil not start anymore
and the fork will be terminated. In the current version of Cherokee
the process would have likely be owned by the user/group of the
cherokee-worker process. If that is truly what the user wanted,
no user or group would have been specified.