Make the SSL tests run
This patch adds support for generating a SnakeOil SSL key during the QA suite
if the system has openssl.  It can take a little while on lower-powered systems
but on anything modern it should barely be noticable.

This means we can cause the SSL tests to be run if the server was built with
OPENSSL support.  In addition we fix the TLS port to be bound to 127.0.0.1
which fixes 073's behaviour in SSL mode on an IPv6 enabled host.

Signed-Off-By: Daniel Silverstone <dsilvers@digital-scurf.org>