MAIN: Do not process special exit codes unless worker is exiting.
When the worker exits, it can use some special exit codes to indicate
to the main process what kind of termination is happening.  This
is then used to control parent behaviour.  Unfortunately if a FastCGI
uses the same codes it can inadvertently kill the parent.

This patch fixes this problem.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>