XSS vulnerability in handler_error via cherokee_buffer_add_escape_html (#1231)
In Issue #1223 LogicalTrust describes an XSS in the 400 page. When providing a buffer that contains a \0 char strpbrk will not be able to find characters afterwards. This leads to the question: should we be able to return escaped sequences containing \0 values. Since the original function assumed those should not exists this code might guard for it. But I do wonder if it is elegant enough to do so. A subsequent question was raised if the 400 page should return "bad input" a discussion on this feature is started in #1230. Additionally a QA-test was added.

Fixes #1223 and #1209 
And resolves CVE-2006-1681.