Fix autogen.sh with recent autoconf-2.71 and automake-1.16 (#1271)
With a recent upgrade of a development system I have noticed the issues from #1249 and #1239. While both did not resolve the problems I encountered I was able to resolve the issues in the following way.

Starting from the work of @rdratlos I noticed that AM_INIT_AUTOMAKE was defined twice, this alone caused the direct breaking of autogen.sh with recent versions of automake and autoconf. Removing the second instance resolved the issue. While @neelpatel05 suggested to use AM_INIT_AUTOMAKE with "foreign subdir-objects" I have still not reproduced that using automake 1.16.5. Hence I have ignored this suggestion, but if someone can show how it breaks without it I'll obviously include it.

autoupdate-2.71 failed apply an update and terminated processing at an old AC_TRY_COMPILE code block of "Check of func and co.." I manually rewrote the AC_COMPILE_IFELSE. Applying autoupdate-2.71 to the rest of the m4 files in the repository allowed me to clean up all the remaining warnings.

Co-authored-by: Thomas Reim <reimth@gmail.com>